# docs/swagger.yaml
definitions:
...
  main.Paginate-main_Showtime:
    properties:
      current_page:
        type: integer
      items:
        items:
          $ref: '#/definitions/main.Showtime'
        type: array
      page_size:
        type: integer
      total_items:
        type: integer
      total_page:
        type: integer
    type: object
...
  main.Response-any:
    properties:
      data: {}
      message:
        type: string
    type: object
  main.Response-array_main_Seat:
    properties:
      data:
        items:
          $ref: '#/definitions/main.Seat'
        type: array
      message:
        type: string
    type: object
...
  main.Response-main_Paginate-main_Showtime:
    properties:
      data:
        $ref: '#/definitions/main.Paginate-main_Showtime'
      message:
        type: string
    type: object
...
  main.Response-main_Showtime:
    properties:
      data:
        $ref: '#/definitions/main.Showtime'
      message:
        type: string
    type: object
...
  main.Seat:
    properties:
      additional_price:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      is_available:
        description: relation
        type: boolean
      name:
        type: string
      room_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  main.Showtime:
    properties:
      available_seat:
        type: integer
      created_at:
        type: string
      end_at:
        type: string
      id:
        type: integer
      movie_id:
        type: integer
      movie_title:
        description: relation
        type: string
      price:
        type: integer
      room_id:
        type: integer
      room_name:
        type: string
      start_at:
        type: string
      total_seat:
        type: integer
      updated_at:
        type: string
    type: object
  main.ShowtimeFilter:
    properties:
      after:
        description: only list showtime after/equal this time
        example: "2006-01-02T15:04:05+08:00"
        type: string
      ids:
        items:
          type: integer
        type: array
      movie_ids:
        items:
          type: integer
        type: array
      room_ids:
        items:
          type: integer
        type: array
    type: object
  main.ShowtimeInput:
    properties:
      end_at:
        example: "2006-01-02T15:05:05+08:00"
        type: string
      movie_id:
        type: integer
      price:
        type: integer
      room_id:
        type: integer
      start_at:
        example: "2006-01-02T15:04:05+08:00"
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/admin/showtimes:
    post:
      consumes:
      - application/json
      description: admin create showtime
      parameters:
      - description: bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: body request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.ShowtimeInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Response-main_Showtime'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Response-any'
      summary: Create Showtime
      tags:
      - schedules
  /api/admin/showtimes/{id}:
    delete:
      consumes:
      - application/json
      description: admin delete showtime by id
      parameters:
      - description: bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: showtime id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Response-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Response-any'
      summary: Delete Showtime
      tags:
      - schedules
    put:
      consumes:
      - application/json
      description: admin update showtime by id
      parameters:
      - description: bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: movie id
        in: path
        name: id
        required: true
        type: integer
      - description: body request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.ShowtimeInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Response-main_Showtime'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Response-any'
      summary: Update Showtime
      tags:
      - schedules
...
  /api/showtimes/{id}:
    get:
      consumes:
      - application/json
      description: get showtime by id
      parameters:
      - description: showtime id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Response-main_Showtime'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Response-any'
      summary: Get Showtime
      tags:
      - schedules
  /api/showtimes/{id}/seats:
    get:
      consumes:
      - application/json
      description: get showtime seats by showtime is
      parameters:
      - description: showtime id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Response-array_main_Seat'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Response-any'
      summary: Get Showtime seats
      tags:
      - schedules
  /api/showtimes/filter:
    post:
      consumes:
      - application/json
      description: filter showtimes
      parameters:
      - description: pagination page
        in: query
        name: page
        type: integer
      - description: pagination page size
        in: query
        name: per_page
        type: integer
      - description: filter
        in: body
        name: request
        schema:
          $ref: '#/definitions/main.ShowtimeFilter'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Response-main_Paginate-main_Showtime'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.Response-any'
      summary: Filter Showtime
      tags:
      - schedules
...
swagger: "2.0"